+++ lib/src/driver/ratel_session.dart
@@ -1,0 +1,5 @@
+import 'query_result.dart';
+
+abstract interface class RatelSession {
+  Future<QueryResult> query(String sql, {Map<String, Object?>? parameters});
+}