+++ lib/src/exceptions.dart
@@ -1,5 +1,0 @@
-import 'package:ratel/ratel.dart' show DatabaseException;
-
-class MappingException extends DatabaseException {
-  const MappingException(super.message);
-}