+++ lib/src/exceptions/database_exception.dart
@@ -1,0 +1,9 @@
+abstract class DatabaseException implements Exception {
+  const DatabaseException(this.message, {this.cause});
+
+  final String message;
+  final Object? cause;
+
+  @override
+  String toString() => '$runtimeType: $message';
+}