+++ lib/src/exceptions/driver_connection_exception.dart
@@ -1,0 +1,5 @@
+import 'database_exception.dart';
+
+class DriverConnectionException extends DatabaseException {
+  const DriverConnectionException(super.message, {super.cause});
+}