+++ lib/src/exceptions/mapping_exception.dart
@@ -1,0 +1,5 @@
+import 'database_exception.dart';
+
+class MappingException extends DatabaseException {
+  const MappingException(super.message, {super.cause});
+}