+++ lib/src/exceptions/query_execution_exception.dart
@@ -1,0 +1,11 @@
+import 'database_exception.dart';
+
+class QueryExecutionException extends DatabaseException {
+  const QueryExecutionException(
+    super.message, {
+    required this.sql,
+    super.cause,
+  });
+
+  final String sql;
+}