+++ lib/src/orm_driver.dart
@@ -1,7 +1,0 @@
-import 'package:ratel/ratel.dart' show RatelDriver;
-
-import 'dialect.dart';
-
-abstract class OrmDriver extends RatelDriver {
-  SqlDialect get dialect;
-}