+++ lib/src/postgres/postgres_session.dart
@@ -1,6 +1,7 @@
 import 'package:postgres/postgres.dart';
-import 'package:ratel/ratel.dart' show QueryResult, RatelSession;
 
+import '../driver/query_result.dart';
+import '../driver/ratel_session.dart';
 import 'query_result_mapper.dart';
 
 class PostgresSession implements RatelSession {