+++ lib/src/postgres/query_result_mapper.dart
@@ -1,5 +1,6 @@
 import 'package:postgres/postgres.dart';
-import 'package:ratel/ratel.dart' show QueryResult;
+
+import '../driver/query_result.dart';
 
 QueryResult toQueryResult(Result result) => QueryResult(
       rows: [for (final row in result) row.toColumnMap()],